Added Network code to JUCE, Added Visual Studio 2017 build (and .gitignore) to JUCE, added Network/State connectivity to MainComponent
-Added Network code to JUCE, contained in Network.h/cpp
-Network Code defines several classes, such as NetworkClient, NetworkServer,  State, and Transmittable
-Inherit and Implement the Transmittable interface to make a class able to be added to the application state
-Synth components are NOT transmittable yet
-Network code works for multiple clients, but management is pretty basic (no locks)
-Added server_code folder to the repo (not JUCE) with an some source to build the standalone server (probably only works on windows, but the source is all there, just not configured for other builds)

FILES ADDED

To JUCE:
     Network.cpp/h
To repository:
     Everything in server_code

FILES MODIFIED

MainComponent.cpp/h

+++ midinote_basic/Source/Synth.cpp
@@ -329,7 +329,7 @@ Synth::Synth()
     releaseLabelX.attachToComponent(&releaseSliderX, false);
 
     updateSettings(frequencySlider.getValue(),
-                   static_cast<Oscillator::WaveType> (floorf(waveSlider.getValue())),
+                   static_cast<Oscillator::WaveType> (static_cast<int>(floorf(waveSlider.getValue()))),
                    dBToVolume(levelSlider.getValue()),
                    panSlider.getValue());
 }
@@ -363,7 +363,7 @@ void Synth::addNote (MidiMessage message)
 
     Oscillator osc (frequencySlider.getValue(),
                     dBToVolume (levelSlider.getValue()),
-                    static_cast<Oscillator::WaveType> (floorf(waveSlider.getValue())));
+                    static_cast<Oscillator::WaveType> (static_cast<int>(floorf(waveSlider.getValue()))));
     std::pair<MidiMessage, Oscillator> val (message, osc);
     int key = message.getNoteNumber();
     currentNotes.insert (std::pair<int, std::pair<MidiMessage, Oscillator>> (key, val));
